Move sdk into zep directory Golang does not allow multiple packages in one directory, so all files under the 'zep-go' directory were moved to 'zep'
